requirements.txt
Flask
flask-cors
Flask-Migrate
flask-sqlalchemy
mysql-connector-python
redis
pydantic
pydantic[email]
python-dotenv
Werkzeug
PyJWT
pika
python-telegram-bot
requests
flasgger
chromadb
langchain
langchain-community
langchain-chroma
